Fixed issue with turned off Autolayout

+++ ObjectiveTrainerApp/Menu/MenuViewController.m
@@ -50,7 +50,7 @@
     return self.menuItems.count;
 }
 
-- (double)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 60.0;
 }
